Mark activity extensions as internal
