+++ src/OpenTelemetry.Api/Internal/SemanticConventions.cs
