+++ src/OpenTelemetry.Api/Internal/SpanAttributeConstants.cs
